test: Unit tests for the new rule's config format
Write tests for rules activations or not
